+++ src/main/java/ec/edu/uce/controller/dto/SubProductoDTO.java
@@ -19,7 +19,7 @@ public class SubProductoDTO {
 
 	private BigDecimal precioVenta;
 
-	private Integer stockActual;
+	private Double stockActual;
 
 	private String proveedor;
 	
@@ -50,7 +50,7 @@ public class SubProductoDTO {
 	 * @param producto
 	 */
 	public SubProductoDTO(Integer id, String codigoBarras, String nombre, String descripcion,
-			Integer cantidadRelacionada, BigDecimal costoPromedio, BigDecimal precioVenta, Integer stockActual,
+			Integer cantidadRelacionada, BigDecimal costoPromedio, BigDecimal precioVenta, Double stockActual,
 			String proveedor, String impuesto, String seccion, String marca, String producto) {
 		super();
 		this.id = id;
@@ -116,11 +116,11 @@ public class SubProductoDTO {
 		this.precioVenta = precioVenta;
 	}
 
-	public Integer getStockActual() {
+	public Double getStockActual() {
 		return stockActual;
 	}
 
-	public void setStockActual(Integer stockActual) {
+	public void setStockActual(Double stockActual) {
 		this.stockActual = stockActual;
 	}
